refactor/navbar
