+++ src/configs/naviteminfo.ts
@@ -1,0 +1,5 @@
+export const navItemsInfo = [
+  { name: 'Home', type: 'link' },
+  { name: 'Articles', type: 'link' },
+  { name: 'Faq', type: 'link' },
+]
